requirements.txt
aider-chat==0.86.2
aiohappyeyeballs==2.6.1
aiohttp>=3.9.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.94.1
anyio>=4.0.0
asgiref>=3.8.0
attrs==26.1.0
backoff==2.2.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.7
click==8.3.1
ConfigArgParse==1.7.1
cryptography==46.0.7
diff-match-patch==20241021
diskcache==5.6.3
distro==1.9.0
Django==5.1.6
django-cors-headers==4.9.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docstring_parser==0.17.0
dotenv==0.9.9
drf-nested-routers==0.95.0
fastapi==0.128.8
fastuuid==0.14.0
filelock==3.20.3
flake8==7.3.0
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.2
google-genai==1.73.0
grep-ast==0.9.0
gunicorn==25.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-aiohttp==0.1.12
huggingface_hub==1.4.1
idna==3.11
importlib_metadata==7.2.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.14.0
json5==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
litellm==1.81.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mixpanel==5.0.0
msgpack==1.1.0
mslex==1.3.0
multidict==6.7.1
networkx==3.4.2
numpy==1.26.4
openai==2.20.0
orjson==3.11.7
oslex==0.1.3
packaging==26.0
pathspec==1.0.4
pexpect==4.9.0
pillow==11.1.0
posthog==7.8.6
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
pydub==0.25.1
pyflakes==3.4.0
Pygments==2.20.0
PyJWT==2.9.0
pypandoc==1.16.2
pyperclip==1.11.0
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2024.1
PyYAML==6.0.3
qrcode==8.0
redis==6.0.0
referencing==0.37.0
regex==2026.1.15
reportlab==4.3.1
requests==2.31.0
rich==15.0.0
rpds-py==0.30.0
scipy==1.15.3
shellingham==1.5.4
shtab==1.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sounddevice==0.5.5
soundfile==0.13.1
soupsieve==2.8.3
sqlparse==0.5.3
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
tree-sitter==0.25.2
tree-sitter-c-sharp==0.23.1
tree-sitter-embedded-template==0.25.0
tree-sitter-language-pack==0.13.0
tree-sitter-yaml==0.7.2
typer==0.23.0
typer-slim==0.23.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.6.3
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
yarl==1.23.0
zipp==3.23.0
